requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
click==8.0.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1 